pyproject.toml
[project]
name = "timebox-evo-client"
version = "0.1.0"
description = "Divoom Timebox EVO controller for macOS via IOBluetooth RFCOMM"
requires-python = ">=3.9"
dependencies = [
    "pyobjc-framework-IOBluetooth",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["timebox"]
